Refactor code with Product Class
Create productClass.js, move Product class from database.js to this. Move creating 6 products from database.js to main.js.
